Renaming and movei simulation commands
